Move non-job function
